Clear selection button & sorting change
the button looks so ugly lmao

Also found & fixed a big bug in the custom sort thing I pushed earlier. Unfortunately, keeping all the old functionality was a lot more complex in this new version, so currently there's only a custom sort order for month. I think this is fine, but let me know if there is another feature you really want to introduce a custom sorting for, then I'll see what I can do.

Finally, made some changes to the data xlsx: everything is now properly capitalized, and the "Other" and "Unknown" categories are now called "~Other" and "~Unknown" so they always appear last in the list ('~' is one of the last ASCII characters)